refactor(main): メインページのコンポーネント分割を行う
