+++ README.md
@@ -3,12 +3,19 @@
 <br>
 
 ## App URL
+
 ### https://env-hub-hazel.vercel.app/
 
-## Qiita記事
-### [【個人開発/学生】".env"をGitHubリポジトリに紐づけて共有出来る"EnvHub"を作成しました。<br>【Next.js + Supabase】](https://qiita.com/warabimochi_26/items/0c86ea1e6dfb84fb1c4a)
+## Qiita 記事
+
+### [【個人開発/学生】".env"を GitHub リポジトリに紐づけて共有出来る"EnvHub"を作成しました。<br>【Next.js + Supabase】](https://qiita.com/warabimochi_26/items/0c86ea1e6dfb84fb1c4a)
 
 ## 主な使用技術
+
 - Next.js(App Router)
 - Supabase
-- 他に使用したライブラリ等はQiita記事をご参照ください。
+- 他に使用したライブラリ等は Qiita 記事をご参照ください。
+
+## V2 からの方針
+
+- PR タイトルが Conventional Commits に沿うようにする
