+++ app/components/toppage/elements/image-in-slider.tsx
@@ -1,0 +1,18 @@
+import Image, { StaticImageData } from "next/image";
+
+interface ImageInSliderProps {
+  ImageSrc: StaticImageData;
+}
+
+export function ImageInSlider({ ImageSrc }: ImageInSliderProps) {
+  return (
+    <div className="relative h-full">
+      <Image
+        src={ImageSrc}
+        alt="EnvHubのスクリーンショット"
+        fill
+        className="object-contain"
+      />
+    </div>
+  );
+}
