+++ app/components/toppage/elements/splide-progress-bar.tsx
@@ -1,0 +1,7 @@
+export function SplideProgressBar() {
+  return (
+    <div className="splide__progress">
+      <div className="splide__progress__bar"></div>
+    </div>
+  );
+}
